Refactoring for curand abstraction replacements. Step 3.
