+++ src/cunumeric/random/randutil/generator_host.cc
@@ -16,6 +16,7 @@
 
 #include "generator.h"
 #include "generator_create.inl"
+#include "cunumeric/random/rnd_aliases.h"
 
 #if !defined(LEGATE_USE_CUDA)
 // the host code of cuRAND try to extern these variables out of nowhere,
@@ -24,29 +25,29 @@ const dim3 blockDim{};
 const uint3 threadIdx{};
 #endif
 
-extern "C" curandStatus_t randutilCreateGeneratorHost(randutilGenerator_t* generator,
-                                                      curandRngType_t rng_type,
-                                                      uint64_t seed,
-                                                      uint64_t generatorID,
-                                                      uint32_t flags)
+extern "C" rnd_status_t randutilCreateGeneratorHost(randutilGenerator_t* generator,
+                                                    randRngType_t rng_type,
+                                                    uint64_t seed,
+                                                    uint64_t generatorID,
+                                                    uint32_t flags)
 {
   return inner_randutilCreateGenerator<randutilimpl::execlocation::HOST>(
     generator, rng_type, seed, generatorID, nullptr);
 }
 
-extern "C" curandStatus_t randutilDestroyGenerator(randutilGenerator_t generator)
+extern "C" rnd_status_t randutilDestroyGenerator(randutilGenerator_t generator)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
   gen->destroy();
   delete gen;
-  return CURAND_STATUS_SUCCESS;
+  return RND_STATUS_SUCCESS;
 }
 
 #pragma region integers
 
 #include "generator_integers.inl"
 
-extern "C" curandStatus_t randutilGenerateIntegers16(
+extern "C" rnd_status_t randutilGenerateIntegers16(
   randutilGenerator_t generator, int16_t* outputPtr, size_t n, int16_t low, int16_t high)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
@@ -56,7 +57,7 @@ extern "C" curandStatus_t randutilGenerateIntegers16(
   return randutilimpl::dispatch<decltype(func), int16_t>(gen, func, n, outputPtr);
 }
 
-extern "C" curandStatus_t randutilGenerateIntegers32(
+extern "C" rnd_status_t randutilGenerateIntegers32(
   randutilGenerator_t generator, int32_t* outputPtr, size_t n, int32_t low, int32_t high)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
@@ -66,7 +67,7 @@ extern "C" curandStatus_t randutilGenerateIntegers32(
   return randutilimpl::dispatch<decltype(func), int32_t>(gen, func, n, outputPtr);
 }
 
-extern "C" curandStatus_t randutilGenerateIntegers64(
+extern "C" rnd_status_t randutilGenerateIntegers64(
   randutilGenerator_t generator, int64_t* outputPtr, size_t n, int64_t low, int64_t high)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
@@ -82,7 +83,7 @@ extern "C" curandStatus_t randutilGenerateIntegers64(
 
 #include "generator_lognormal.inl"
 
-extern "C" curandStatus_t randutilGenerateLogNormalEx(
+extern "C" rnd_status_t randutilGenerateLogNormalEx(
   randutilGenerator_t generator, float* outputPtr, size_t n, float mean, float stddev)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
@@ -92,7 +93,7 @@ extern "C" curandStatus_t randutilGenerateLogNormalEx(
   return randutilimpl::dispatch<decltype(func), float>(gen, func, n, outputPtr);
 }
 
-extern "C" curandStatus_t randutilGenerateLogNormalDoubleEx(
+extern "C" rnd_status_t randutilGenerateLogNormalDoubleEx(
   randutilGenerator_t generator, double* outputPtr, size_t n, double mean, double stddev)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
@@ -108,7 +109,7 @@ extern "C" curandStatus_t randutilGenerateLogNormalDoubleEx(
 
 #include "generator_normal.inl"
 
-extern "C" curandStatus_t randutilGenerateNormalEx(
+extern "C" rnd_status_t randutilGenerateNormalEx(
   randutilGenerator_t generator, float* outputPtr, size_t n, float mean, float stddev)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
@@ -118,7 +119,7 @@ extern "C" curandStatus_t randutilGenerateNormalEx(
   return randutilimpl::dispatch<decltype(func), float>(gen, func, n, outputPtr);
 }
 
-extern "C" curandStatus_t randutilGenerateNormalDoubleEx(
+extern "C" rnd_status_t randutilGenerateNormalDoubleEx(
   randutilGenerator_t generator, double* outputPtr, size_t n, double mean, double stddev)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
@@ -134,10 +135,10 @@ extern "C" curandStatus_t randutilGenerateNormalDoubleEx(
 
 #include "generator_poisson.inl"
 
-extern "C" curandStatus_t randutilGeneratePoissonEx(randutilGenerator_t generator,
-                                                    uint32_t* outputPtr,
-                                                    size_t n,
-                                                    double lambda)
+extern "C" rnd_status_t randutilGeneratePoissonEx(randutilGenerator_t generator,
+                                                  uint32_t* outputPtr,
+                                                  size_t n,
+                                                  double lambda)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
   poisson func;
@@ -151,9 +152,9 @@ extern "C" curandStatus_t randutilGeneratePoissonEx(randutilGenerator_t generato
 
 #include "generator_raw.inl"
 
-extern "C" curandStatus_t randutilGenerateRawUInt32(randutilGenerator_t generator,
-                                                    uint32_t* outputPtr,
-                                                    size_t n)
+extern "C" rnd_status_t randutilGenerateRawUInt32(randutilGenerator_t generator,
+                                                  uint32_t* outputPtr,
+                                                  size_t n)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
   raw<uint32_t> func;
@@ -166,7 +167,7 @@ extern "C" curandStatus_t randutilGenerateRawUInt32(randutilGenerator_t generato
 
 #include "generator_uniform.inl"
 
-extern "C" curandStatus_t randutilGenerateUniformEx(
+extern "C" rnd_status_t randutilGenerateUniformEx(
   randutilGenerator_t generator, float* outputPtr, size_t n, float low, float high)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
@@ -177,7 +178,7 @@ extern "C" curandStatus_t randutilGenerateUniformEx(
   return randutilimpl::dispatch<decltype(func), float>(gen, func, n, outputPtr);
 }
 
-extern "C" curandStatus_t randutilGenerateUniformDoubleEx(
+extern "C" rnd_status_t randutilGenerateUniformDoubleEx(
   randutilGenerator_t generator, double* outputPtr, size_t n, double low, double high)
 {
   randutilimpl::basegenerator* gen = (randutilimpl::basegenerator*)generator;
