+++ src/cunumeric/random/rnd_types.h
@@ -17,10 +17,7 @@
 #pragma once
 
 #ifdef USE_STL_RANDOM_ENGINE_
-using rnd_status_t = int;
-enum class randRngType : int { STL_MT19937 = 1 };
-using randRngType_t              = randRngType;
-constexpr int RND_STATUS_SUCCESS = 0;
+#include "cunumeric/random/rnd_aliases.h"
 
 #define CHECK_RND_ENGINE(expr)                             \
   do {                                                     \
@@ -55,10 +52,10 @@ static inline randRngType get_rndRngType(cunumeric::BitGeneratorType kind)
 
 #else
 #include "cunumeric/random/curand_help.h"
-using rnd_status_t                        = curandStatus_t;
-using randRngType                         = curandRngType;
-using randRngType_t                       = curandRngType_t;
-constexpr rnd_status_t RND_STATUS_SUCCESS = CURAND_STATUS_SUCCESS;
+
+// in this case include aliases _after_ curand_help
+//
+#include "cunumeric/random/rnd_aliases.h"
 
 #define CHECK_RND_ENGINE(expr) CHECK_CURAND((expr))
 #define randutil_check_status randutil_check_curand
